backend/requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
nltk==3.9.4
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
vaderSentiment==3.3.2
Werkzeug==3.1.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
